turned modules to functions
