+++ js/main.js
@@ -1,6 +1,21 @@
 import * as dom from 'dom-helpers';
-import welcomeElement from 'template-welcome';
+import makeElement from 'template-welcome';
+
+/**
+ * welcome page data structure
+ * @const
+ * @type {object}
+ */
+const welcome = {
+  gameName: 'Угадай мелодию',
+  content: {
+    title: 'Правила игры',
+    text: `Правила просты&nbsp;— за&nbsp;2 минуты дать максимальное количество правильных ответов.
+          На&nbsp;каждую мелодию всего 3 варианта ответа. Удачи!`
+  },
+  playButton: 'Начать игру'
+};
 
 window.onload = () => {
-  dom.renderElement(welcomeElement);
+  dom.renderElement(makeElement(welcome));
 };
